Rename the core program layer
