+++ debug/modelPrograms/voictents/serialized/sanity-snapshot/output-file-digest-list.yaml
@@ -19,36 +19,36 @@
   - "fileName": "render-type-script-file-relationships"
     "digest": "a6e3e44a3f41e11d3a20a7c8defe5c8a24d4f774"
   - "fileName": "test-build-add-metadata-for-serialization"
-    "digest": "46414fe34ab5948440d345a460778282a23283a0"
+    "digest": "34fcae362a69ada035263af2217aa445abee7202"
   - "fileName": "test-cached-on-disk-datum"
-    "digest": "8d3e2ab8a33ad48740e123f7dcd092351c3f1f02"
+    "digest": "280fd5a911a17207edeac33199173e3a003a0646"
   - "fileName": "test-error-serialization"
-    "digest": "6557775b075b4339c65c02f37e3ac5b2f39c5130"
+    "digest": "68f2cec198db1c52c774e1ef1d2acbb02d6c30a8"
   - "fileName": "test-estinant-error"
-    "digest": "dd33e84f947b80615d0939d8701a1c2e1af50ad6"
+    "digest": "e243fea1c63c4551456bac007006307fdc1c5ba6"
   - "fileName": "test-get-custom-typed-datum"
-    "digest": "dc7a6ec2aa837c4c33ab1ae7bbc84e8927d2cfd4"
+    "digest": "d04442865cec8319c3718f724fc22621faac75db"
   - "fileName": "test-get-type-script-typed-datum"
-    "digest": "4fb0d867642b4ab2a255f59d2e780ba489aa843f"
+    "digest": "10d8509f7a920416495d2b6a11b00607e532b565"
   - "fileName": "test-graph-render"
     "digest": "f6baaf641e6fc0f75a910bb49e02943ec72adf36"
   - "fileName": "test-joining-one-to-many"
-    "digest": "dc68e3a989876d142549add006ddd047c39a8475"
+    "digest": "a9e671be3fbdf35561f7f7c5c2ad7ed717b0a7bc"
   - "fileName": "test-joining-one-to-one"
-    "digest": "68753f7e4101e76b2d63e9e185f66e79231acc1b"
+    "digest": "eb742bd23e44ed3c1f3ebdb3cdb4c26ad2754ddb"
   - "fileName": "test-joining-one-to-voictent"
-    "digest": "e6cf908fc1868cfff8d5c460ecad170a64018d1e"
+    "digest": "ef49517e18cce3707bca6d78f78cb1227bf3a755"
   - "fileName": "test-joining-voictent-to-voictent"
-    "digest": "2fe63281250ec50040367012d2c9d863daceac89"
+    "digest": "a3c7539b507eaecbcd27c7d764bfa938fc65ab65"
   - "fileName": "test-json-serialization"
-    "digest": "6dafea3ea8ec42f2cdb9cecafd9fd66271f2b70a"
+    "digest": "03a8afda65d40662bb5c6af3c2afabc4b60793f0"
   - "fileName": "test-releasing-a-left-voictent-multiple-times"
-    "digest": "10c66cfb0beb057106671449bfe6c922bcde8357"
+    "digest": "bfd74a7ca6b67a8621a75b1002955d0729765195"
   - "fileName": "test-releasing-a-right-voictent-multiple-times"
-    "digest": "91d7bb5c0333df30004240add791c027af93e0e9"
+    "digest": "b36efb002e6de3ce33d3b5fe357aee83d2bec70a"
   - "fileName": "test-serialize"
-    "digest": "029ae24314ce003bf1aa386c90b87de9c9342c1b"
+    "digest": "13ee71c3b7e35208b1d5e6b764c4a479e552199c"
   - "fileName": "test-untriggered-cology-error"
-    "digest": "69d592fa1bf0228579e047a3795303ca49c8bfb1"
+    "digest": "bdd6747b7dc32924d2e9d0e809d565714cb58d9e"
   - "fileName": "test-voictent-input"
-    "digest": "db189fb1e469cf1042a0910d15d6bb7a56204fe6"
+    "digest": "995b3fe2a6be5aff47a2ea7faef869031c3e079a"
