+++ packages/voictents-and-estinants-engine/ci.sh
@@ -5,14 +5,14 @@ set -e
 ## test-json-serialization
 echo "# test-json-serialization"
 echo "Testing JsonSerializableCollection"
-npx ts-node packages/voictents-and-estinants-engine/src/example-programs/testJsonSerialization.ts
+npx ts-node packages/voictents-and-estinants-engine/src/core-programs/testJsonSerialization.ts
 bash checkUncommittedDebug.sh
 printf "\n"
 
 ## test-error-serialization
 echo "# test-error-serialization"
 echo "Testing ErrorSerializableCollection"
-npx ts-node packages/voictents-and-estinants-engine/src/example-programs/testErrorSerialization.ts
+npx ts-node packages/voictents-and-estinants-engine/src/core-programs/testErrorSerialization.ts
 bash checkUncommittedDebug.sh
 printf "\n"
 
@@ -21,70 +21,70 @@ printf "\n"
 ## test-build-add-metadata-for-serialization
 echo "# test-build-add-metadata-for-serialization"
 echo "Testing consuming each item in a collection, and "buildAddMetadataForSerialization""
-npx ts-node packages/voictents-and-estinants-engine/src/example-programs/testBuildAddMetadataForSerialization.ts
+npx ts-node packages/voictents-and-estinants-engine/src/core-programs/testBuildAddMetadataForSerialization.ts
 bash checkUncommittedDebug.sh
 printf "\n"
 
 ## test-estinant-error
 echo "# test-estinant-error"
 echo "Testing that the engine forwards errors to an error collection"
-npx ts-node packages/voictents-and-estinants-engine/src/example-programs/engine-behavior/testEstinantError.ts
+npx ts-node packages/voictents-and-estinants-engine/src/core-programs/engine-behavior/testEstinantError.ts
 bash checkUncommittedDebug.sh
 printf "\n"
 
 ## test-voictent-input
 echo "# test-voictent-input"
 echo "Testing consuming a collection as a whole"
-npx ts-node packages/voictents-and-estinants-engine/src/example-programs/engine-behavior/testVoictentInput.ts
+npx ts-node packages/voictents-and-estinants-engine/src/core-programs/engine-behavior/testVoictentInput.ts
 bash checkUncommittedDebug.sh
 printf "\n"
 
 ## test-joining-one-to-one
 echo "# test-joining-one-to-one"
 echo "Testing joining each item in one collection to one item from another collection"
-npx ts-node packages/voictents-and-estinants-engine/src/example-programs/engine-behavior/testJoiningOneToOne.ts
+npx ts-node packages/voictents-and-estinants-engine/src/core-programs/engine-behavior/testJoiningOneToOne.ts
 bash checkUncommittedDebug.sh
 printf "\n"
 
 ## test-joining-one-to-many
 echo "# test-joining-one-to-many"
 echo "Testing joining each item in one collection to multiple items from another"
-npx ts-node packages/voictents-and-estinants-engine/src/example-programs/engine-behavior/testJoiningOneToMany.ts
+npx ts-node packages/voictents-and-estinants-engine/src/core-programs/engine-behavior/testJoiningOneToMany.ts
 bash checkUncommittedDebug.sh
 printf "\n"
 
 ## test-joining-one-to-voictent
 echo "# test-joining-one-to-voictent"
 echo "Testing joining each item in one collection the an entire different collection as a whole"
-npx ts-node packages/voictents-and-estinants-engine/src/example-programs/engine-behavior/testJoiningOneToVoictent.ts
+npx ts-node packages/voictents-and-estinants-engine/src/core-programs/engine-behavior/testJoiningOneToVoictent.ts
 bash checkUncommittedDebug.sh
 printf "\n"
 
 ## test-joining-voictent-to-voictent
 echo "# test-joining-voictent-to-voictent"
 echo "Testing joining one collection as a whole to another collection as a whole"
-npx ts-node packages/voictents-and-estinants-engine/src/example-programs/engine-behavior/testJoiningVoictentToVoictent.ts
+npx ts-node packages/voictents-and-estinants-engine/src/core-programs/engine-behavior/testJoiningVoictentToVoictent.ts
 bash checkUncommittedDebug.sh
 printf "\n"
 
 ## test-releasing-a-left-voictent-multiple-times
 echo "# test-releasing-a-left-voictent-multiple-times"
 echo "Testing a left collection that stops accumulating items for one engine tick"
-npx ts-node packages/voictents-and-estinants-engine/src/example-programs/engine-behavior/testReleasingALeftVoictentMultipleTimes.ts
+npx ts-node packages/voictents-and-estinants-engine/src/core-programs/engine-behavior/testReleasingALeftVoictentMultipleTimes.ts
 bash checkUncommittedDebug.sh
 printf "\n"
 
 ## test-releasing-a-right-voictent-multiple-times
 echo "# test-releasing-a-right-voictent-multiple-times"
 echo "Testing a right collection that stops accumulating items for one engine tick"
-npx ts-node packages/voictents-and-estinants-engine/src/example-programs/engine-behavior/testReleasingARightVoictentMultipleTimes.ts
+npx ts-node packages/voictents-and-estinants-engine/src/core-programs/engine-behavior/testReleasingARightVoictentMultipleTimes.ts
 bash checkUncommittedDebug.sh
 printf "\n"
 
 ## test-untriggered-cology-error
 echo "# test-untriggered-cology-error"
 echo "Testing that the engine emits an error when a cology is left untriggered"
-npx ts-node packages/voictents-and-estinants-engine/src/example-programs/engine-behavior/testUntriggeredCologyError.ts
+npx ts-node packages/voictents-and-estinants-engine/src/core-programs/engine-behavior/testUntriggeredCologyError.ts
 bash checkUncommittedDebug.sh
 printf "\n"
 
@@ -93,21 +93,21 @@ printf "\n"
 ## test-get-type-script-typed-datum
 echo "# test-get-type-script-typed-datum"
 echo "Testing "getTypeScriptTypedDatum""
-npx ts-node packages/voictents-and-estinants-engine/src/example-programs/testGetTypeScriptTypedDatum.ts
+npx ts-node packages/voictents-and-estinants-engine/src/core-programs/testGetTypeScriptTypedDatum.ts
 bash checkUncommittedDebug.sh
 printf "\n"
 
 ## test-get-custom-typed-datum
 echo "# test-get-custom-typed-datum"
 echo "Testing "getCustomTypedDatum""
-npx ts-node packages/voictents-and-estinants-engine/src/example-programs/testGetCustomTypedDatum.ts
+npx ts-node packages/voictents-and-estinants-engine/src/core-programs/testGetCustomTypedDatum.ts
 bash checkUncommittedDebug.sh
 printf "\n"
 
 ## test-serialize
 echo "# test-serialize"
 echo "Testing "serialize""
-npx ts-node packages/voictents-and-estinants-engine/src/example-programs/testSerialize.ts
+npx ts-node packages/voictents-and-estinants-engine/src/core-programs/testSerialize.ts
 bash checkUncommittedDebug.sh
 printf "\n"
 
@@ -188,7 +188,7 @@ printf "\n"
 ## test-cached-on-disk-datum
 echo "# test-cached-on-disk-datum"
 echo "Perfoming test run of "testCachedOnDiskDatum""
-npx ts-node packages/voictents-and-estinants-engine/src/example-programs/testCachedOnDiskDatum.ts
+npx ts-node packages/voictents-and-estinants-engine/src/core-programs/testCachedOnDiskDatum.ts
 bash checkUncommittedDebug.sh
 printf "\n"
