+++ packages/voictents-and-estinants-engine/src/adapted-programs/programmable-units/in-memory-cache/zornable.ts
@@ -1,6 +1,6 @@
 import { Merge } from 'type-fest';
 import { Voque } from '../../../core/engine/voque';
-import { AbstractSerializableIndexByName } from '../../../example-programs/abstractSerializableVoictent';
+import { AbstractSerializableIndexByName } from '../../../core-programs/abstractSerializableVoictent';
 import { Gepp } from '../../../core/engine-shell/voictent/gepp';
 import { Hubblepup } from '../../../core/engine-shell/quirm/hubblepup';
