+++ packages/voictents-and-estinants-engine/src/adapted-programs/programmable-units/type-script-file/canonical-declaration/canonicalDeclarationOmissionList.ts
@@ -41,22 +41,22 @@ export const canonicalDeclarationOmissionList: LintAssertionOmission[] = [
   'packages/voictents-and-estinants-engine/src/adapted-programs/programs/scaffold-file/assembleScaffoldedFile.ts',
   'packages/voictents-and-estinants-engine/src/adapted-programs/programs/scaffold-file/types.ts',
   'packages/voictents-and-estinants-engine/src/adapted-programs/programs/test-graph-render/testGraphRender.ts',
-  'packages/voictents-and-estinants-engine/src/example-programs/engine-behavior/testEstinantError.ts',
-  'packages/voictents-and-estinants-engine/src/example-programs/engine-behavior/testJoiningOneToMany.ts',
-  'packages/voictents-and-estinants-engine/src/example-programs/engine-behavior/testJoiningOneToOne.ts',
-  'packages/voictents-and-estinants-engine/src/example-programs/engine-behavior/testJoiningOneToVoictent.ts',
-  'packages/voictents-and-estinants-engine/src/example-programs/engine-behavior/testJoiningVoictentToVoictent.ts',
-  'packages/voictents-and-estinants-engine/src/example-programs/engine-behavior/testReleasingALeftVoictentMultipleTimes.ts',
-  'packages/voictents-and-estinants-engine/src/example-programs/engine-behavior/testReleasingARightVoictentMultipleTimes.ts',
-  'packages/voictents-and-estinants-engine/src/example-programs/engine-behavior/testUntriggeredCologyError.ts',
-  'packages/voictents-and-estinants-engine/src/example-programs/engine-behavior/testVoictentInput.ts',
-  'packages/voictents-and-estinants-engine/src/example-programs/testBuildAddMetadataForSerialization.ts',
-  'packages/voictents-and-estinants-engine/src/example-programs/testCachedOnDiskDatum.ts',
-  'packages/voictents-and-estinants-engine/src/example-programs/testErrorSerialization.ts',
-  'packages/voictents-and-estinants-engine/src/example-programs/testGetCustomTypedDatum.ts',
-  'packages/voictents-and-estinants-engine/src/example-programs/testGetTypeScriptTypedDatum.ts',
-  'packages/voictents-and-estinants-engine/src/example-programs/testJsonSerialization.ts',
-  'packages/voictents-and-estinants-engine/src/example-programs/testSerialize.ts',
+  'packages/voictents-and-estinants-engine/src/core-programs/engine-behavior/testEstinantError.ts',
+  'packages/voictents-and-estinants-engine/src/core-programs/engine-behavior/testJoiningOneToMany.ts',
+  'packages/voictents-and-estinants-engine/src/core-programs/engine-behavior/testJoiningOneToOne.ts',
+  'packages/voictents-and-estinants-engine/src/core-programs/engine-behavior/testJoiningOneToVoictent.ts',
+  'packages/voictents-and-estinants-engine/src/core-programs/engine-behavior/testJoiningVoictentToVoictent.ts',
+  'packages/voictents-and-estinants-engine/src/core-programs/engine-behavior/testReleasingALeftVoictentMultipleTimes.ts',
+  'packages/voictents-and-estinants-engine/src/core-programs/engine-behavior/testReleasingARightVoictentMultipleTimes.ts',
+  'packages/voictents-and-estinants-engine/src/core-programs/engine-behavior/testUntriggeredCologyError.ts',
+  'packages/voictents-and-estinants-engine/src/core-programs/engine-behavior/testVoictentInput.ts',
+  'packages/voictents-and-estinants-engine/src/core-programs/testBuildAddMetadataForSerialization.ts',
+  'packages/voictents-and-estinants-engine/src/core-programs/testCachedOnDiskDatum.ts',
+  'packages/voictents-and-estinants-engine/src/core-programs/testErrorSerialization.ts',
+  'packages/voictents-and-estinants-engine/src/core-programs/testGetCustomTypedDatum.ts',
+  'packages/voictents-and-estinants-engine/src/core-programs/testGetTypeScriptTypedDatum.ts',
+  'packages/voictents-and-estinants-engine/src/core-programs/testJsonSerialization.ts',
+  'packages/voictents-and-estinants-engine/src/core-programs/testSerialize.ts',
   'packages/voictents-and-estinants-engine/src/utilities/constructor-function/constructorFunctionBuilder.ts',
   'packages/voictents-and-estinants-engine/src/utilities/constructor-function/namedConstructorFunctionBuilder.ts',
   'packages/voictents-and-estinants-engine/src/utilities/constructor-function/types.ts',
