+++ packages/voictents-and-estinants-engine/src/adapter/digikikify.ts
@@ -39,9 +39,9 @@ import {
 } from '../core/engine/voictent2';
 import { GenericVoque } from '../core/engine/voque';
 import { ProgramErrorGepp } from '../adapted-programs/programmable-units/error/programError';
-import { GenericAbstractSerializableSourceVoque } from '../example-programs/abstractSerializableVoictent';
-import { buildAddMetadataForSerialization } from '../example-programs/buildAddMetadataForSerialization';
-import { SerializableVoictent } from '../example-programs/serializableVoictent';
+import { GenericAbstractSerializableSourceVoque } from '../core-programs/abstractSerializableVoictent';
+import { buildAddMetadataForSerialization } from '../core-programs/buildAddMetadataForSerialization';
+import { SerializableVoictent } from '../core-programs/serializableVoictent';
 import { ProgramFileCache } from '../utilities/programFileCache';
 import {
   FileSystemNodeVoictent,
