+++ packages/voictents-and-estinants-engine/src/core-programs/abstractSerializableVoictent.ts
