+++ packages/voictents-and-estinants-engine/src/core-programs/buildAddMetadataForSerialization.ts
