+++ packages/voictents-and-estinants-engine/src/core-programs/engine-behavior/serializableErrorVoictent.ts
