+++ packages/voictents-and-estinants-engine/src/core-programs/jsonSerializableVoictent.ts
