+++ packages/voictents-and-estinants-engine/src/core-programs/serializableVoictent.ts
