+++ packages/voictents-and-estinants-engine/src/core-programs/testGetTypeScriptTypedDatum.ts
